python/requirements-audit.txt
pandas==2.3.2
polars==1.36.1
pyarrow==21.0.0
duckdb==1.4.0
PyYAML==6.0.3
